Converted version of the dataset
The dataset is converted with the help of the dataset_cleaner. At some rows the data is not filled as some people have used the app on different times. All the rows have two indeces with person and day. The mood is the average of the mood the people have filled in on the day. IF they haven't filled in the mood the average is 0.